add jsonp for get requests from the client, fixes browserify usage
