+++ package.json
@@ -23,7 +23,9 @@
   },
   "main": "./index",
   "dependencies": {
-    "request": "2.12.0"
+    "request": "2.12.0",
+    "jsonp": "0.0.4",
+    "browser-request": "^0.3.2"
   },
   "devDependencies": {
     "mocha": "*",
